requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-option-group==0.5.7
Deprecated==1.2.18
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotty-dict==1.3.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.3.5
python-gitlab==5.6.0
python-semantic-release==9.21.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
shellingham==1.5.4
smmap==5.0.2
sqlparse==0.5.3
tomlkit==0.13.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wrapt==1.17.2
